requirements.txt
openai
fastapi
uvicorn
python-dotenv
pypdf
numpy
streamlit